backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
fastf1>=3.5.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
plotly>=5.20.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
joblib>=1.3.0
python-dotenv>=1.0.0
httpx>=0.27.0
